Add files via upload
The Open Images Dataset is a collection of millions of images annotated with labels from a hierarchical ontology of thousands of object categories. The first code piece provided in this project utilizes the Open Images Dataset API to download images of various beverages, both alcoholic and non-alcoholic, for the purpose of training a machine-learning model to classify these images. The notebook provides step-by-step instructions on how to access the API, authenticate, and download the images into appropriate directories based on their class labels. This notebook serves as the first step in preparing the data for machine learning model training.